changed return type of strlcat

+++ ft_strlcat.c
@@ -6,14 +6,14 @@
 /*   By: igngonza <igngonza@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/13 14:13:09 by igngonza          #+#    #+#             */
-/*   Updated: 2024/03/19 11:03:46 by igngonza         ###   ########.fr       */
+/*   Updated: 2024/04/16 18:10:54 by igngonza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 #include <string.h>
 
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, char *src, unsigned int size)
 {
 	unsigned int	i;
 	unsigned int	j;
